Code-freezed/mpai-cae-arp/pyproject.toml
[tool.poetry]
name = "mpai-cae-arp"
version = "0.5.0"
description = "The MPAI CAE-ARP software API"
authors = ["Matteo Spanio <dev2@audioinnova.com>"]
readme = "README.md"
packages = [{include = "mpai_cae_arp"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "~1.24.4"   # because of numba (from librosa) requirement <=1.24
pydantic = "^2.3.0"
pyyaml = "^6.0"
librosa = "^0.10.1"
grpcio-tools = "^1.53.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"
yapf = "^0.40.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.1"
toml = "^0.10.2"

[tool.poetry.group.docs.dependencies]
sphinx = "^6.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -n auto -W error::RuntimeWarning"
testpaths = ["tests"]

[tool.coverage.run]
relative_files = true

[tool.yapf]
blank_line_before_nested_class_or_def = true
column_limit = 88

[tool.pylint]
max-line-length = 88
extension-pkg-whitelist=['pydantic']
disable = [
    "C0103",  # Invalid name
    "C0114",  # Missing module docstring
    "C0115",  # Missing class docstring
    "C0116",  # Missing function or method docstring
    "C0301",  # Line too long
    "E1101",  # Module has no member
]
